Chore: Add issue templates and config file (#2)
* chore(issue-templates): add issue template for bug_reports

* chore(issue-template): add issue template for feature requests

* chore(issue-template): add issue template for documentation issues

* chore(issue-template): add config file for issue templates

* fix(bug-report): typo

* fix(issue-template-config): typo

* fix(issue-template-config): delete discord server section because it does not exist

The section on a Discord Server people can join is deleted. However, it might be introduced later. And then, this section is re-introduced.

* feat(issue-template-bug-report): change rendering from python to kotlin

* feat(issue-template-bug-report): add current app version to options

An issuer needs to select the CityKey Android App version the issue relates to. This commit adds the current app version 1.3.15 to the options of versions that can be chosen.

* feat(issue-template-bug-report): add system information issuers must provide

Issuers have to provide information on the system the bug was encountered.
This commit adds the device model and and android version issuers have to provide.

* feat(issue-template-bug-report): add section on development environment

Issuers shall provide as much information as possible about the bug such that we can identify the issue faster.
This commit adds a section on the development environment, where we ask for the kotlin version, the build tools version, and the operationg system among others.

* feat(bug-report-system-environment): add manufacturer and device architecture

This commit adds the manufacturer and device architecture to the system environment information issuers should provide.

* feat(system-environment): add kotlin code that logs system environment information

This commit adds a Kotlin code snipped that logs the system environment information, which is convenient for the issuer.

* chore(bug-report): specify programming language in code segment.

* fix(bug-report): change to Kotlin comments