[CI] Add CODEOWNERS file (#11)
* ci(CODEOWNERS): add CODEOWNERS file

This commit adds the first version of the CODEOWNERS file, which is used to automatically assign reviewers to pull requests.

* ci(CODEOWNERS): change global code owners reference in an attempt to fix error

This commit changes the global code owners reference in an attempt to fix the error.

* ci(CODEOWNERS): remove teams subdomain from global code owners reference in an attempt to fix error

This commit changes the global code owners reference in an attempt to fix the error.